Update bot to use Artieficer classes
Full explanation in commit 808b9fa of the Artieficer repository

In short, some functionalities implemented in this bot could be useful
in all future bots that use the Artieficer framework, so they were
rewritten there.

One positive of this is that people who use Artieficer don't have to
rewrite boilerplate constantly in each of their bot instances, but it
could also limit their options.